requirements.txt
advent-of-code-data==2.0.3
aocd-example-parser==2023.12.20
beautifulsoup4==4.12.3
networkx==3.4.2
numpy==2.1.3
Pebble==5.0.7
PuLP==3.3.0
shapely==2.1.2
soupsieve==2.6
tqdm==4.67.1
urllib3==2.2.3
